server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain>=0.2.0
langchain-mistralai>=1.0.0
mistralai>=1.0.0
requests==2.31.0
python-json-logger==2.0.7
google-genai>=0.3.0
elevenlabs>=0.2.24
